Add support for dynamic themes
Allows users to create their own themes via /themes folder. For creation refer to the  readme.

Fix a few bugs. Notifications not appearing, app arbitrarily crashing.

I've stored some code on my pc that contains emoji support. I will not be pushing this to the main commit due to a streamline issue I'm having. Snapchat defines the textbox users type in as a div, so I can't reset the cursor when I apply the emojis to the chat. If people have demand for this I will gladly release it under a secondary build/branch.